refactor(shared): inline the single-consumer devAssertionMode module into dispatcher (SCHX-5)
dispatcher.ts was isDevAssertionMode's only importer; the module's docstring still claimed two. The env check becomes a private function there, keeping the browser-safety rationale.
